[#26] 댓글 조회, 작성, 수정, 삭제 작업
- 기존 테이블에 빠진 필드 추가
- 댓글 조회, 작성, 수정, 삭제 작업

+++ src/main/java/com/example/kfanboy/board/service/BoardService.java
@@ -71,9 +71,13 @@ public class BoardService {
 	}
 
 	@Transactional
-	public void delete(final BoardDeleteRequestDto boardDeleteRequestDto) {
+	public void delete(final Long memberId, final BoardDeleteRequestDto boardDeleteRequestDto) {
 		Board board = boardRepository.findById(boardDeleteRequestDto.boardId())
 			.orElseThrow(() -> new CustomException(ErrorMessage.BOARD_NOT_FOUND));
+
+		if (!Objects.equals(board.getMemberId(), memberId)) {
+			throw new CustomException(ErrorMessage.BOARD_WRITER_NOT_MATCHED);
+		}
 
 		boardRepository.delete(board);
 	}
